modelnumber added

+++ controllers/categoryController.js
@@ -54,13 +54,23 @@ const updateCategory = asyncHandler(async (req, res) => {
     res.json(error);
   }
 });
+
 const deleteCategory = asyncHandler(async (req, res) => {
   const { id } = req.params;
   try {
+    
+    const category = await Category.findById(id);
+    if (!category) {
+      return res.status(404).json({ message: "Category not found" });
+    }
+
     await Category.findByIdAndDelete(id);
+
     res.status(200).json({ message: "Category deleted" });
   } catch (error) {
-    res.json(error);
+    console.error(error);
+    res.status(500).json({ message: "Failed to delete category" });
   }
 });
+
 module.exports = { createCategory, getAllCategory, getCategory, updateCategory, deleteCategory };
